METAMAC-1457: [CODES] Órdenes. Orden alfabético, cada vez que cambian los códigos
git-svn-id: http://svn.arte-consultores.com/svn/arte/istac/metamac/metamac-srm/trunk/metamac-srm@44107 05f23f4b-1f0b-0410-ba1b-ad9509270e59

+++ metamac-srm-core/src/test/java/org/siemac/metamac/srm/core/facade/serviceapi/SrmCoreServiceFacadeCodesTest.java
@@ -1238,55 +1238,55 @@ public class SrmCoreServiceFacadeCodesTest extends SrmBaseTest {
             // Code 01
             CodeMetamacVisualisationResult code = getCodeMetamacVisualisationResult(codes, CODELIST_1_V2_CODE_1);
             assertEquals("Name codelist-1-v2-code-1", code.getName());
-            assertEquals(Long.valueOf(1), code.getOrder());
+            assertEquals(Integer.valueOf(1), code.getOrder());
         }
         {
             // Code 02
             CodeMetamacVisualisationResult code = getCodeMetamacVisualisationResult(codes, CODELIST_1_V2_CODE_2);
             assertEquals(null, code.getName());
-            assertEquals(Long.valueOf(2), code.getOrder());
+            assertEquals(Integer.valueOf(2), code.getOrder());
         }
         {
             // Code 02 01
             CodeMetamacVisualisationResult code = getCodeMetamacVisualisationResult(codes, CODELIST_1_V2_CODE_2_1);
             assertEquals(null, code.getName());
-            assertEquals(Long.valueOf(1), code.getOrder());
+            assertEquals(Integer.valueOf(1), code.getOrder());
         }
         {
             // Code 02 01 01
             CodeMetamacVisualisationResult code = getCodeMetamacVisualisationResult(codes, CODELIST_1_V2_CODE_2_1_1);
             assertEquals(null, code.getName());
-            assertEquals(Long.valueOf(0), code.getOrder());
+            assertEquals(Integer.valueOf(0), code.getOrder());
         }
         {
             // Code 02 02
             CodeMetamacVisualisationResult code = getCodeMetamacVisualisationResult(codes, CODELIST_1_V2_CODE_2_2);
             assertEquals(null, code.getName()); // it has not name
-            assertEquals(Long.valueOf(0), code.getOrder());
+            assertEquals(Integer.valueOf(0), code.getOrder());
         }
         {
             // Code 03
             CodeMetamacVisualisationResult code = getCodeMetamacVisualisationResult(codes, CODELIST_1_V2_CODE_3);
             assertEquals("name code-3", code.getName());
-            assertEquals(Long.valueOf(0), code.getOrder());
+            assertEquals(Integer.valueOf(0), code.getOrder());
         }
         {
             // Code 04
             CodeMetamacVisualisationResult code = getCodeMetamacVisualisationResult(codes, CODELIST_1_V2_CODE_4);
             assertEquals(null, code.getName());
-            assertEquals(Long.valueOf(3), code.getOrder());
+            assertEquals(Integer.valueOf(3), code.getOrder());
         }
         {
             // Code 04 01
             CodeMetamacVisualisationResult code = getCodeMetamacVisualisationResult(codes, CODELIST_1_V2_CODE_4_1);
             assertEquals(null, code.getName());
-            assertEquals(Long.valueOf(0), code.getOrder());
+            assertEquals(Integer.valueOf(0), code.getOrder());
         }
         {
             // Code 04 01 01
             CodeMetamacVisualisationResult code = getCodeMetamacVisualisationResult(codes, CODELIST_1_V2_CODE_4_1_1);
             assertEquals("Name codelist-1-v2-code-4-1-1", code.getName());
-            assertEquals(Long.valueOf(0), code.getOrder());
+            assertEquals(Integer.valueOf(0), code.getOrder());
         }
     }
 
@@ -2488,10 +2488,10 @@ public class SrmCoreServiceFacadeCodesTest extends SrmBaseTest {
         {
             List<CodeMetamacVisualisationResult> codes = srmCoreServiceFacade.retrieveCodesByCodelistUrn(getServiceContextAdministrador(), codelistUrn, "es", visualisationUrn);
             assertEquals(9, codes.size());
-            assertEquals(Long.valueOf(0), getCodeMetamacVisualisationResult(codes, CODELIST_1_V2_CODE_1).getOrder());
-            assertEquals(Long.valueOf(1), getCodeMetamacVisualisationResult(codes, CODELIST_1_V2_CODE_2).getOrder());
-            assertEquals(Long.valueOf(2), getCodeMetamacVisualisationResult(codes, CODELIST_1_V2_CODE_3).getOrder());
-            assertEquals(Long.valueOf(3), getCodeMetamacVisualisationResult(codes, CODELIST_1_V2_CODE_4).getOrder());
+            assertEquals(Integer.valueOf(0), getCodeMetamacVisualisationResult(codes, CODELIST_1_V2_CODE_1).getOrder());
+            assertEquals(Integer.valueOf(1), getCodeMetamacVisualisationResult(codes, CODELIST_1_V2_CODE_2).getOrder());
+            assertEquals(Integer.valueOf(2), getCodeMetamacVisualisationResult(codes, CODELIST_1_V2_CODE_3).getOrder());
+            assertEquals(Integer.valueOf(3), getCodeMetamacVisualisationResult(codes, CODELIST_1_V2_CODE_4).getOrder());
         }
         // Update
         srmCoreServiceFacade.updateCodeInOrderVisualisation(getServiceContextAdministrador(), codeUrn, visualisationUrn, Integer.valueOf(2));
@@ -2499,10 +2499,10 @@ public class SrmCoreServiceFacadeCodesTest extends SrmBaseTest {
         {
             List<CodeMetamacVisualisationResult> codes = srmCoreServiceFacade.retrieveCodesByCodelistUrn(getServiceContextAdministrador(), codelistUrn, "es", visualisationUrn);
             assertEquals(9, codes.size());
-            assertEquals(Long.valueOf(0), getCodeMetamacVisualisationResult(codes, CODELIST_1_V2_CODE_2).getOrder());
-            assertEquals(Long.valueOf(1), getCodeMetamacVisualisationResult(codes, CODELIST_1_V2_CODE_3).getOrder());
-            assertEquals(Long.valueOf(2), getCodeMetamacVisualisationResult(codes, CODELIST_1_V2_CODE_1).getOrder());
-            assertEquals(Long.valueOf(3), getCodeMetamacVisualisationResult(codes, CODELIST_1_V2_CODE_4).getOrder());
+            assertEquals(Integer.valueOf(0), getCodeMetamacVisualisationResult(codes, CODELIST_1_V2_CODE_2).getOrder());
+            assertEquals(Integer.valueOf(1), getCodeMetamacVisualisationResult(codes, CODELIST_1_V2_CODE_3).getOrder());
+            assertEquals(Integer.valueOf(2), getCodeMetamacVisualisationResult(codes, CODELIST_1_V2_CODE_1).getOrder());
+            assertEquals(Integer.valueOf(3), getCodeMetamacVisualisationResult(codes, CODELIST_1_V2_CODE_4).getOrder());
         }
     }
